requirements.txt
numpy
opencv-python
Pillow
torch==2.5.1+cu118
torchvision==0.20.1+cu118
torchaudio==2.5.1+cu118
tqdm
huggingface-hub
line_profiler
memory_profiler
tensorflow
tensorflow_datasets